archive/old/device/requirements.txt
python-socketio~=5.8.0